Refine Pipeline data structure
Move bluocean types into a separated file
